+++ Source/SwiftLintFramework/Extensions/SwiftSyntax/String+Extensions.swift
@@ -1,0 +1,6 @@
+extension String {
+    /// Trims all whitespace from a string
+    var trimmed: String {
+        trimmingCharacters(in: .whitespacesAndNewlines)
+    }
+}