+++ Source/SwiftLintFramework/Models/ObjectType.swift
@@ -1,0 +1,6 @@
+/// ObjectTypes are types that can conform to protocols or inherit from a parent class.
+public enum ObjectType: String, CaseIterable {
+    case `class`
+    case `enum`
+    case `struct`
+}